+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - `add_scenarios!` method for `Vector{Pair{Symbol, Scenario}}` 
 - store parameters in `Simulation`
 - `OrderedDict` in `Platform`
+- new `plot` results
+- fix bug with single `saveat`
+- fix bugs in saving callback
+- global sensitivity for `MCResult`
+- change headers for parameters table for fitting 
 
 ## 0.4.1